software_and_ai/ai_predictive_models/requirements.txt
paho-mqtt==1.6.1
numpy==1.26.4
scipy==1.12.0
scikit-learn==1.4.1.post1
